restructure java files to match similar format as FireBase
