+++ android/src/main/java/io/fullstack/firestack/FirestackInstanceIdService.java
@@ -1,16 +1,15 @@
 package io.fullstack.firestack;
 
-/**
- * Created by nori on 2016/09/12.
- */
-import android.content.Intent;
-import android.os.Bundle;
 import android.util.Log;
+import android.os.Bundle;
+import android.content.Intent;
 
 import com.google.firebase.iid.FirebaseInstanceId;
 import com.google.firebase.iid.FirebaseInstanceIdService;
 
-public class InstanceIdService extends FirebaseInstanceIdService {
+import io.fullstack.firestack.messaging.FirestackMessaging;
+
+public class FirestackInstanceIdService extends FirebaseInstanceIdService {
 
     private static final String TAG = "FSInstanceIdService";
 
@@ -21,7 +20,7 @@ public class InstanceIdService extends FirebaseInstanceIdService {
     public void onTokenRefresh() {
         String refreshedToken = FirebaseInstanceId.getInstance().getToken();
         Log.d(TAG, "Refreshed token: " + refreshedToken);
-        Intent i = new Intent(Messaging.INTENT_NAME_TOKEN);
+        Intent i = new Intent(FirestackMessaging.INTENT_NAME_TOKEN);
         Bundle bundle = new Bundle();
         bundle.putString("token", refreshedToken);
         i.putExtras(bundle);
