+++ android/src/main/java/io/fullstack/firestack/FirestackPackage.java
@@ -6,11 +6,18 @@ import com.facebook.react.ReactPackage;
 import com.facebook.react.bridge.JavaScriptModule;
 import com.facebook.react.bridge.NativeModule;
 import com.facebook.react.bridge.ReactApplicationContext;
+import com.facebook.react.uimanager.UIManagerModule;
 import com.facebook.react.uimanager.ViewManager;
 
 import java.util.List;
 import java.util.ArrayList;
 import java.util.Collections;
+
+import io.fullstack.firestack.auth.FirestackAuth;
+import io.fullstack.firestack.storage.FirestackStorage;
+import io.fullstack.firestack.database.FirestackDatabase;
+import io.fullstack.firestack.analytics.FirestackAnalytics;
+import io.fullstack.firestack.messaging.FirestackMessaging;
 
 @SuppressWarnings("unused")
 public class FirestackPackage implements ReactPackage {
@@ -25,12 +32,12 @@ public class FirestackPackage implements ReactPackage {
     @Override
     public List<NativeModule> createNativeModules(ReactApplicationContext reactContext) {
         List<NativeModule> modules = new ArrayList<>();
-        modules.add(new Module(reactContext, reactContext.getBaseContext()));
-        modules.add(new Auth(reactContext));
-        modules.add(new Database(reactContext));
-        modules.add(new Analytics(reactContext));
-        modules.add(new Storage(reactContext));
-        modules.add(new Messaging(reactContext));
+        modules.add(new FirestackModule(reactContext, reactContext.getBaseContext()));
+        modules.add(new FirestackAuth(reactContext));
+        modules.add(new FirestackDatabase(reactContext));
+        modules.add(new FirestackAnalytics(reactContext));
+        modules.add(new FirestackStorage(reactContext));
+        modules.add(new FirestackMessaging(reactContext));
         return modules;
     }
