+++ android/src/main/java/io/fullstack/firestack/Utils.java
@@ -37,9 +37,7 @@ public class Utils {
   /**
    * send a JS event
    **/
-  public static void sendEvent(final ReactContext context,
-                               final String eventName,
-                               final WritableMap params) {
+  public static void sendEvent(final ReactContext context, final String eventName, final WritableMap params) {
     if (context.hasActiveCatalystInstance()) {
       context
           .getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class)
